\encoding{UTF-8}
\name{smoothHR-package}
\alias{smoothHR-package}
\docType{package}
\title{Smooth Hazard Ratio Curves taking a Reference Value}
\description{
	Provides flexible hazard ratio curves allowing non-linear relationships between continuous predictors and survival.
	To better understand the effects that each continuous covariate has on the outcome, results are expressed in terms of hazard ratio curves,
	taking a specific covariate value as reference. Confidence bands for these curves are also derived.
}

\details{
	\tabular{ll}{
		Package: \tab smoothHR\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2013-03-28\cr
		License: \tab GPL (>= 2)\cr
		LazyLoad: \tab yes\cr
		LazyData: \tab yes\cr
	}
}

\author{
	Artur Agostinho Araújo and Luís Meira-Machado \email{lmachado@math.uminho.pt}\cr
	Maintainer: Artur Agostinho Araújo \email{artur.stat@gmail.com}
}

\references{
	Carmen Cadarso-Suarez, Luis Meira-Machado, Thomas Kneib and Francisco Gude. Flexible hazard ratio curves for continuous predictors in multi-state models: a P-spline approach. Statistical Modelling, 2010, 10:291-314.
}
