\encoding{UTF-8}
\name{whas500}
\alias{whas500}
\docType{data}
\title{Worcester Heart Attack Study WHAS500 Data}
\description{Data from the Worcester Heart Attack Study}
\usage{data(whas500)}

\format{A data frame with 500 observations with 22 variables.}

\details{
Data from the Worcester Heart Attack Study whose main goal was to describe factors associated
with trends over time in the incidence and survival rates following
hospital admission for acute myocardial infarction.
}

\source{
Worcester Heart Attack Study data from Dr. Robert J. Goldberg of
the Department of Cardiology at the University of Massachusetts Medical
School.
}

\references{
Hosmer, D. W. and Lemeshow, S. and May, S. (2008). \emph{Applied Survival Analysis: Regression Modeling of Time to Event Data: Second Edition}, John Wiley and Sons Inc., New York, NY.
}

\examples{
data(whas500)
}

\keyword{datasets}
