\encoding{UTF-8}
\name{smoothHR-package}
\alias{smoothHR-package}
\docType{package}
\title{Smooth Hazard Ratio Curves Taking a Reference Value}
\description{
	Provides flexible hazard ratio curves allowing non-linear relationships between continuous predictors and survival.
	To better understand the effects that each continuous covariate has on the outcome, results are expressed in terms of hazard ratio curves,
	taking a specific covariate value as reference. Confidence bands for these curves are also derived.
}

\details{
	\tabular{ll}{
		Package: \tab smoothHR\cr
		Type: \tab Package\cr
		Version: \tab 1.0.2\cr
		Date: \tab 2015-10-29\cr
		License: \tab GPL (>= 2)\cr
		LazyLoad: \tab yes\cr
		LazyData: \tab yes\cr
	}
}

\author{
	Artur Araújo and Luís Meira-Machado \email{lmachado@math.uminho.pt}\cr
	Maintainer: Artur Araújo \email{artur.stat@gmail.com}
}

\references{
	Carmen Cadarso-Suarez, Luis Meira-Machado, Thomas Kneib and Francisco Gude. Flexible hazard ratio curves for continuous predictors in multi-state models: a P-spline approach. Statistical Modelling, 2010, 10:291-314.

	Eilers, Paul H. and Marx, Brian D. (1996). Flexible smoothing with B-splines and penalties. Statistical Science, 11, 89-121.

	Hosmer, D.W. and Lemeshow, S. and May, S. (2008) Applied Survival Analysis: Regression Modeling of Time to Event Data: Second Edition, John Wiley and Sons Inc., New York, NY.

	Hurvich, C.M. and Simonoff, J.S. and Tsai, Chih-Ling (1998). Smoothing parameter selection in nonparametric regression using an improved Akaike information criterion, JRSSB, volume 60, 271–293.
}

\keyword{package}
